Improve usage/input validation and logging, support screenlock PIN.
This changeset dramatically improves the validation around usage of this
library, logging errors and/or warnings when invalid actions are performed.

Additionally, this change eliminates the hard requirement to call login()
introduced in PR #15 and adds support for thermostats with the screenlock PIN
enabled.

Finally, this changeset addresses a few minor stylistical issues,
such as trailing whitespace, consistent string quoting and double slashes
in generated URLs as well as increments the version to 0.10