Fix handling of firmware version for T3900
I have a T3900 which reports a firmware version of 2.19.16, which breaks
the version parsing that is currently done, which treats the version
number as a single floating point number.  Change to handling the
firmwave version as a tuple of ints, which provides the lexicographic
ordering needed for the current version checking.